Refact donation widget custom finish message configuration page #297
