+++ app/modules/widgets/__plugins__/donation/components/index.js
@@ -1,2 +1,3 @@
 export { default } from './__donation__'
 export { default as SettingsMenu } from './settings-menu'
+export { default as DonationTellAFriend } from './donation-tell-a-friend'
