+++ app/modules/widgets/__plugins__/donation/pages/index.js
@@ -1,2 +1,2 @@
 export { default as SettingsDonationPage } from './settings-donation-page'
-export { default as SettingsDonationFinishPage } from './settings-donation-finish-page'
+export { default as SettingsFinishMessagePage } from './settings-finish-message-page'
