+++ app/modules/widgets/components/form-finish-message.scss.js
@@ -1,0 +1,20 @@
+export const editor = ({
+  borderRadius: 3,
+  backgroundColor: '#fff',
+  border: '1px solid #efefef'
+})
+
+export const editorFocus = ({
+  border: '1px solid #51a7e8',
+  outline: 'none',
+  boxShadow: 'inset 0 1px 2px rgba(0,0,0,0.075), 0 0 5px rgba(81,167,232,0.5)',
+  top: 58
+})
+
+export const editorToolbar = ({
+  left: 0
+})
+
+export const editorContainer = ({
+  minHeight: 315
+})
