+++ app/modules/widgets/components/index.js
@@ -3,4 +3,5 @@ export { default as BlockTag } from './block-tag'
 export { default as InputTag } from './input-tag'
 export { default as Tag } from './tag'
 export { default as FormFooter } from './form-footer'
+export { default as FormFinishMessage } from './form-finish-message'
 export { default as WidgetOverlay } from './widget-overlay'
